requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==25.10.2
Automat==25.4.16
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==46.0.2
daphne==4.2.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
hiredis==3.3.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.2
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.2
pyOpenSSL==25.3.0
python-dotenv==1.2.1
python3-openid==3.2.0
redis==7.0.1
requests==2.32.5
requests-oauthlib==2.0.0
service-identity==24.2.0
social-auth-app-django==5.5.1
social-auth-core==4.8.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zope.interface==8.1
daphne==4.0.0  